Removed StaticBuffer and DynamicBuffer from PicoTree.
