+++ src/pico_tree/pico_tree/internal/memory.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <vector>
+#include <type_traits>
 
 namespace pico_tree {
 
@@ -101,49 +101,6 @@ class ListPool {
   Chunk* end_;
   //! \brief Index within the last chunk.
   std::size_t index_;
-};
-
-//! \brief Static MemoryBuffer using a vector. It is a simple memory buffer
-//! making deletions of recursive elements a bit easier.
-//! \details The buffer owns all memory returned by Allocate() and all memory is
-//! released when the buffer is destroyed.
-template <typename T>
-class StaticBuffer {
- public:
-  //! \brief Type allocated and stored by the buffer.
-  using ValueType = T;
-
-  //! Creates a StaticBuffer having space for \p size elements.
-  inline explicit StaticBuffer(std::size_t const size) {
-    buffer_.reserve(size);
-  }
-
-  //! \brief Creates an item and returns a pointer to it.
-  template <typename... Args>
-  inline T* Allocate(Args&&... args) {
-    buffer_.emplace_back(std::forward<Args>(args)...);
-    return &buffer_.back();
-  }
-
- private:
-  //! \private
-  std::vector<T> buffer_;
-};
-
-//! \brief Dynamic MemoryBuffer using a ListPool.
-//! \details The buffer owns all memory returned by Allocate() and all memory is
-//! released when the buffer is destroyed.
-template <typename T, std::size_t ChunkSize = 256>
-class DynamicBuffer : public ListPool<T, ChunkSize> {
- public:
-  //! \brief Type allocated and stored by the buffer.
-  using ValueType = typename ListPool<T, ChunkSize>::ValueType;
-
-  //! \brief Creates a DynamicBuffer.
-  inline DynamicBuffer() = default;
-  //! \brief Creates a DynamicBuffer. Ignores the argument in favor of a common
-  //! interface with the StaticBuffer.
-  inline explicit DynamicBuffer(std::size_t const) {}
 };
 
 }  // namespace internal
